refactoring formatters into a new folder
