+++ x-pack/legacy/plugins/apm/public/utils/formatters/__test__/size.test.ts
@@ -3,60 +3,9 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import {
-  asPercent,
-  asDuration,
-  getFixedByteFormatter,
-  asDynamicBytes
-} from '../duration_formatters';
+import { getFixedByteFormatter, asDynamicBytes } from '../size';
 
-describe('duration formatters', () => {
-  describe('asDuration', () => {
-    it('formats correctly with defaults', () => {
-      expect(asDuration(null)).toEqual('N/A');
-      expect(asDuration(undefined)).toEqual('N/A');
-      expect(asDuration(0)).toEqual('0 μs');
-      expect(asDuration(1)).toEqual('1 μs');
-      expect(asDuration(1000)).toEqual('1,000 μs');
-      expect(asDuration(1000 * 1000)).toEqual('1,000 ms');
-      expect(asDuration(1000 * 1000 * 10)).toEqual('10,000 ms');
-      expect(asDuration(1000 * 1000 * 20)).toEqual('20.0 s');
-      expect(asDuration(60000000 * 10)).toEqual('10.0 min');
-      expect(asDuration(3600000000 * 1.5)).toEqual('1.5 h');
-    });
-
-    it('formats without unit', () => {
-      expect(asDuration(1000, { withUnit: false })).toEqual('1,000');
-    });
-
-    it('falls back to default value', () => {
-      expect(asDuration(undefined, { defaultValue: 'nope' })).toEqual('nope');
-    });
-  });
-
-  describe('asPercent', () => {
-    it('should divide and format item as percent', () => {
-      expect(asPercent(3725, 10000, 'n/a')).toEqual('37.3%');
-    });
-
-    it('should format when numerator is 0', () => {
-      expect(asPercent(0, 1, 'n/a')).toEqual('0.0%');
-    });
-
-    it('should return fallback when denominator is undefined', () => {
-      expect(asPercent(3725, undefined, 'n/a')).toEqual('n/a');
-    });
-
-    it('should return fallback when denominator is 0 ', () => {
-      expect(asPercent(3725, 0, 'n/a')).toEqual('n/a');
-    });
-
-    it('should return fallback when numerator or denominator is NaN', () => {
-      expect(asPercent(3725, NaN, 'n/a')).toEqual('n/a');
-      expect(asPercent(NaN, 10000, 'n/a')).toEqual('n/a');
-    });
-  });
-
+describe('size formatters', () => {
   describe('byte formatting', () => {
     const bytes = 10;
     const kb = 1000 + 1;
